feat(stage-pages): sync speech module page with doubao provider settings
Extract the Doubao voice parameters into a shared voice-settings-panel
component rendered by both the provider settings page and the speech
module page, so an edit on one page appears on the other. The module
page writes its model selection back to config.resourceId and adopts
the configured resourceId when the Doubao provider becomes active.

Hide the generic Pitch slider and the SSML toggle for providers without
SSML support, and move the page's hardcoded English strings into the
i18n package (en + zh-Hans).
